fix: relocate secure-memory primitives to a neutral SecureMemory namespace
PinnedPoolArray<T>, PinnedPoolArrayList, CountedEqualityComparer<T>, and
ICountedEqualityComparer<T> lived under
SecretSharingDotNet.Cryptography.SecureArray, which classified them as
cryptography-layer types. They are foundational secure-memory primitives with
no math or cryptography dependency, yet the math layer depends on them deeply
(SecureBigInteger stores its limbs in a PinnedPoolArray<ulong>). The old
namespace therefore forced the math layer to depend upward on the cryptography
layer.

Move the four types to a new top-level namespace,
SecretSharingDotNet.SecureMemory (src/SecureMemory/), so both layers depend
downward on a shared primitive. The change is namespace-only; no type, member,
or behaviour change. Consumers replace
`using SecretSharingDotNet.Cryptography.SecureArray;` with
`using SecretSharingDotNet.SecureMemory;`. README examples and the Demo.Console
sample are updated, and the tests are mirrored to tests/SecureMemory/.

All six test target frameworks stay green (1706 net8/9/10, 1692 net4x) with no
test-count change.
