+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a domain exception hierarchy rooted at `SecretSharingException` (derives from `Exception`): `InvalidShareException` for a malformed share and `ReconstructionException` for a set of individually valid shares that cannot be reconstructed. Catching `SecretSharingException` handles any Shamir-domain failure uniformly.
 
 ### Changed
+- Moved the secure-memory primitives (`PinnedPoolArray<T>`, `PinnedPoolArrayList`, `CountedEqualityComparer<T>`, `ICountedEqualityComparer<T>`) from `SecretSharingDotNet.Cryptography.SecureArray` to a new top-level `SecretSharingDotNet.SecureMemory` namespace — they are foundational primitives shared by the math and cryptography layers, not cryptography-specific types. Replace `using SecretSharingDotNet.Cryptography.SecureArray;` with `using SecretSharingDotNet.SecureMemory;`.
 - Parsing or constructing a `Share<TNumber>` from its serialized form (the pinned-char constructor and the byte-array constructor) now throws `InvalidShareException` for a malformed share, instead of a mix of `ArgumentException`, `FormatException`, and `ArgumentOutOfRangeException`. `InvalidShareException` derives from `SecretSharingException` (not from the former BCL types), so `catch` clauses targeting `FormatException` / `ArgumentException` for parse failures must be updated. Passing an already-decoded, non-positive index through the `(index, value)` calculator constructor still throws `ArgumentOutOfRangeException`.
 - `SecretReconstructor<TNumber>.Reconstruction` now throws `ReconstructionException` (duplicate share indices, no maximum y-value, or a zero / non-invertible denominator during interpolation) instead of `ArgumentException`. `ReconstructionException` derives from `SecretSharingException` (not from `ArgumentException`), so `catch (ArgumentException)` around reconstruction must be updated.
 - Updated `Microsoft.SourceLink.GitHub` from `10.0.300` to `10.0.301` (build-time dependency; moves the transitive `System.IO.Hashing` from `10.0.8` to `10.0.10`).
