+++ src/Math/Numerics/SecureBigInteger.cs
@@ -33,7 +33,7 @@
 
 namespace SecretSharingDotNet.Math.Numerics;
 
-using Cryptography.SecureArray;
+using SecureMemory;
 using System;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
