+++ src/SecureMemory/PinnedPoolArrayList.cs
@@ -29,7 +29,7 @@
 // THE SOFTWARE.
 #endregion
 
-namespace SecretSharingDotNet.Cryptography.SecureArray;
+namespace SecretSharingDotNet.SecureMemory;
 
 using Extension;
 using System;
