+++ src/SecureMemory/PinnedPoolArray`1.cs
@@ -29,7 +29,7 @@
 // THE SOFTWARE.
 #endregion
 
-namespace SecretSharingDotNet.Cryptography.SecureArray;
+namespace SecretSharingDotNet.SecureMemory;
 
 using System;
 using System.Buffers;
