+++ tests/Cryptography/SecureBigInteger/SecretTest.cs
@@ -32,7 +32,7 @@
 namespace SecretSharingDotNetTest.Cryptography.SecureBigInteger;
 
 using SecretSharingDotNet.Cryptography;
-using SecretSharingDotNet.Cryptography.SecureArray;
+using SecretSharingDotNet.SecureMemory;
 using SecretSharingDotNet.Cryptography.SecureInput;
 using SecretSharingDotNet.Math;
 using SecretSharingDotNet.Math.Numerics;
